+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KN1QEWCABR7PYS1B8GHBWB2R"
+        "01KN1X5WV14J3BBQT514P824WZ"
     )
     r"""The unique identifier of the data source"""