+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KN1QEWC8XN8YBK3HB2YK0C2V"
+        "01KN1X5WTZ64GZDSAP1WYE7SPF"
     )
     r"""The unique identifier of the data source"""