+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KN1QEWCEGH1Z0G85KNZV4SER"
+        "01KN1X5WV36WXC5V2HT0Y0Q6A1"
     )
     r"""The unique identifier of the data source"""