+++ packages/orq-rc/RELEASES.md
@@ -6749,3 +6749,13 @@ Based on:
 - [python v4.7.0-rc.34] packages/orq-rc
 ### Releases
 - [PyPI v4.7.0-rc.34] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.34 - packages/orq-rc
+
+## 2026-03-31 12:19:24
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.0 (2.879.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.7.0-rc.35] packages/orq-rc
+### Releases
+- [PyPI v4.7.0-rc.35] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.35 - packages/orq-rc